requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
idna==3.6
numpy==1.26.4
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
xlsx2csv==0.8.2
